requirements.txt
flask
ipykernel
google-generativeai
PyMuPDF
python-dotenv